+++ src/extension.ts
@@ -3,6 +3,7 @@ import { FileReader } from "./classes/Util/FileReader";
 import { CommandRegistrator } from "./classes/Util/CommandRegistrator";
 import { CompletionItemRegistrator } from "./classes/Util/CompletionItemRegistrator";
 import { DefinitionProviderRegistrator } from "./classes/Util/DefinitionProviderRegistrator";
+// import { SignatureHelpRegistrator } from "./classes/Util/SignatureHelpRegistrator";
 import { FileWatcher } from "./classes/Util/FileWatcher";
 
 export async function activate(context: vscode.ExtensionContext) {
@@ -27,6 +28,8 @@ export async function activate(context: vscode.ExtensionContext) {
 					CommandRegistrator.register(context);
 
 					DefinitionProviderRegistrator.register(context);
+
+					// SignatureHelpRegistrator.register(context);
 
 					resolve();
 				} catch (error) {